requirements.txt
setuptools
build
python-magic